Implement the PRE_DEFINED_REF in PumpProbeProcessor
- Rename LaserOnOffProcessor to PumpProbeProcessor;
- Change threshold mask from float to int;
- Add one more normalizer ROI_SUB (ROI1 - ROI2) for normalization;
- Default threshold mask is changed from -1e5 to 1e5. For JungFrau
  data, we found that many pixels are in negative values;
- More unittest

+++ karaboFAI/gui/plot_widgets/image_view.py
@@ -337,7 +337,7 @@ class ImageAnalysis(ImageView):
         self._image_data.update()
         self.setImage(self._image_data.masked_mean)
 
-    @QtCore.pyqtSlot(float, float)
+    @QtCore.pyqtSlot(int, int)
     def onThresholdMaskChange(self, lb, ub):
         if self._image_data is None:
             return
